requirements.txt
flask
twilio
openai
json
duffel
mysql.connector